+++ index.js
@@ -78,6 +78,7 @@ const allRules = {
   'no-unknown-property': require('./lib/rules/no-unknown-property'),
   'no-unsafe': require('./lib/rules/no-unsafe'),
   'no-unstable-nested-components': require('./lib/rules/no-unstable-nested-components'),
+  'no-unused-class-component-methods': require('./lib/rules/no-unused-class-component-methods'),
   'no-unused-prop-types': require('./lib/rules/no-unused-prop-types'),
   'no-unused-state': require('./lib/rules/no-unused-state'),
   'no-will-update-set-state': require('./lib/rules/no-will-update-set-state'),